Fix compilation of metadata
For some reason, even though the code itself compiles and works,
metadata couldn't compile. This is fixed by more carefully
separating which code is compiled for each target, so unneeded
code doesn't get included in the Native Android implementation.
