--- core/androidNative/test/Util.kt
@@ -1,10 +1,0 @@
-/*
- * Copyright 2019-2024 JetBrains s.r.o. and contributors.
- * Use of this source code is governed by the Apache 2.0 License that can be found in the LICENSE.txt file.
- */
-
-package kotlinx.datetime.test
-
-import kotlin.test.Ignore
-
-actual typealias NoAndroid = Ignore
